fixed problem with async api get randomWord
I was using useEffect to fetch the first word, but I also needed one when the user clicked to start the game so the user could restart it. This situation was making the app to render twice to start the game. Fixed by removing the fetch from the useEffect and now when the randomWord is filled the user goes from loading to the game